# fix #79 - allow empty array in ArrayEach