--- a/.travis.yml
+++ b/.travis.yml
@@ -1,3 +1,3 @@
 language: go
-go: 1.6
+go: 1.7
 script: go test -v ./.