@buger 2017-01-30
Thank you! Can you fix the tests?