@vkd 2017-01-30
Wrong author in commits, will be reopened